Find/Replace overlay: move components into internal package
moves helper classes for the find/replace overlay into an internal
package.
